Merged in KL-3579-introduce-system-age-ranges (pull request #176)
[KL-3579] Introduce system age ranges and some refactor

* [KL-3579] ♻️  Create test helper for common functionality on tests
Mocha in inspired in rspec and this is a pretty common concept on it.
This commit adds a test helper for running common functionality for all
tests. Also since this are related with opening connections and doing
operations on the DB, I have only added the helper to integration tests.
This way I don't slow down unit tests with something currently doesn't
need. Finally I have seen we open too many connections for doing
operations this is something we should refactor further on.

* [KL-3579] 🧹 Refactor functions for creating system entities
Pending the change to run the application initialisers on start up since
I need to configure it in the deployment pipeline.

* [KL-3579] 👨‍✈️  Add system default age ranges initializer

Approved-by: Matthew Carey