Merged in KL-4521-fix-bug-upload-user-csv (pull request #240)
[KL-4521] 🔧 Adapt users CSV upload to multiusers

* [KL-4521] 🔧 Adapt users CSV upload to multiusers
This commit changes the users CSV upload to not rely anymore on email or
phone and create new users when necessary. Aligning with what multiusers
should do.

* [KL-4521] 🔩 Increase the acceptable CSV file size
This size is the max size we are uploading currently. We might need to
revisit this number

Approved by me, which is not ideal but since is blocker for Chrysalis on Monday. This will be deployed to Alpha first, so any errors we can address them on UK time without delaying us. Any improvement will be a fast follow up.