Merged in KL-4520-fix-permission-bug (pull request #239)
[KL-4520] 🐞 Fix permission bug

* [KL-4520] 🐞 Fix permission bug
This centralises the logic of allowance in one class and makes sure that
super admins are considered when checking allowance

* [KL-4520] 🔧 Use userPermissions for check allowance on school memb
This centralises the logic of allowance in one class and makes sure that
super admins are considered when checking allowance

Approved by me, which is not ideal but since is blocker for Chrysalis on Monday. This will be deployed to Alpha first, so any errors we can address them on UK time without delaying us. Any improvement will be a fast follow up.