Get one working as test
Change error message for duplicated child to show the parent which is already in the system, not the parent from the row which is being added.
Move to custom errors.
Build the exact error message to compare with the result of the test rather than testing each one by one in an error prone way. It uses the messages from our customErrors file so changes to message formatting shouldn't fail tests, but the wrong message or location should.
None of the existing tests tested anything apart from that it was rejected.
Add missing duplicate child message, plus the default CSV error so we can full move to customErrors.