Merged in testing/csv-fix-with-example (pull request #457)
Testing/csv fix with example

* If a CSV Error, Convert the graphql error response back to our custom error
and keep all the individual CSV validations so the test can check them.

Currently nothing more detailed than "ERR_CSV_BAD_INPUT" can actually
be tested as the reason for rejection.

* gitignore

* Test is meant to check multiple orgs added for single user.
We start by creating an org for our user, then run one in via csv.
Currently they have the same name, so the error is actually for
uploading duplicate orgs rather than the user belonging to two seperate
orgs.

* Get one working as test

Change error message for duplicated child to show the parent which is already in the system, not the parent from the row which is being added.
Move to custom errors.
Build the exact error message to compare with the result of the test rather than testing each one by one in an error prone way. It uses the messages from our customErrors file so changes to message formatting shouldn't fail tests, but the wrong message or location should.
None of the existing tests tested anything apart from that it was rejected.
Add missing duplicate child message, plus the default CSV error so we can full move to customErrors.
* Fix multi org for user lookup


* Revert business logic changes to be safe


* Revert to use old error constants


* Remove unneeded import which fails lint


* Merged master into testing/csv-fix-with-example

Approved-by: Oliver Harrison
Approved-by: Matthew Carey