Merged in improve-linting-and-formatting (pull request #476)
Improve linting and formatting

* Don't run eslint with --fix in CI

The --fix option means that where possible eslint will fix linting errors and save them to the file system.
But `npm run lint` in called in CI, which means 1) the fixes won't be saved anyway 2) we wouldn't want them to be, as the changes are invisibile to a PR reviewer

https://eslint.org/docs/user-guide/command-line-interface#--fix

* Fix prettier format command and add CI check for formatting

Because the pattern being used for prettier was not quoted, we were not formatting all files.

The reason why unquoted globs don't work on mac is pretty funny/interesting
https://medium.com/@jakubsynowiec/you-should-always-quote-your-globs-in-npm-scripts-621887a2a784

```
NPM lifecycle uses `sh -s` to run commands on OS X/Linux
...
OS X ships with bash-3.2 but globstar (** → globstar matches zero or more directories and subdirectories) is supported starting from Bash version 4.0.
```

And sure enough:
```
richardsommerville~$sh -c 'echo ./**/*.{ts,test.ts}'
./src/app.ts ./src/main.dev.ts ./src/main.ts ./src/model.ts ./src/token.ts ./tests/makeDevToken.ts ./tests/testHelper.ts ./**/*.test.ts
```

Any code not in that list was not getting formated


Approved-by: Oliver Harrison
Approved-by: Max Nicholson
Approved-by: Enrique M Figuerola Gomez