Merged in UD-764-valid-tokens-on-all-endpoints (pull request #469)
UD-764 valid tokens on all endpoints

* [UD-764] Only allow calmid-debug issued tokens in development mode

We only allow this if the enviroment is explictly "development", because its possible production deployments might set it to values we do not anticipate (alpha, beta, etc).
The assumption is that it's better to accidently break dev enviroments then to accidently allow dev tokens in production.

* [UD-764] Throw errors when token check fails authentication

This breaks lots of tests for endpoints that where not already enforcing authentication
For these endpoints, the client must now supply a token, but it can still be any valid token.
To highlight this, where tokens had to be added to tests new users with no test-specific permissions are used.
There should probably be follow up work to audit what authorisation checks these endpoints should be doing.


Approved-by: Oliver Harrison
Approved-by: Enrique M Figuerola Gomez