Merged in UD-900-remove-filters-on-child-nodes (pull request #483)
UD-900 remove filters on child nodes

* UD-900 Optimize handling of class connection results

We need to return classes in the same order they were requested
However this might not match the order they are returned by postgres

The previous code used `Array.find` to do this but that requires searching through the search arrary for each class.
Now we only search that array once to build a map, and then use that to look up each class.

* UD-900 Remove unnecessary joins in class connections

The class IDs supplied by the loader come from the parent class of the entity we're fetching
https://bitbucket.org/calmisland/kidsloop-user-service/src/9dae70d93f0d3a366558b057b4fc3bdc83a3ef1f/src/schemas/class.ts#lines-165
These have already been filtered by model.classConnections
https://bitbucket.org/calmisland/kidsloop-user-service/src/9dae70d93f0d3a366558b057b4fc3bdc83a3ef1f/src/schemas/class.ts#lines-233
https://bitbucket.org/calmisland/kidsloop-user-service/src/9dae70d93f0d3a366558b057b4fc3bdc83a3ef1f/src/model.ts#lines-746

Because all the joins/filters removed in this commit are examining the already-filter class, they aren't needed.

We've also made the decision to not filter on the attributes of child nodes even where it would make a difference to the query results
https://kidsloop.slack.com/archives/C01T8EZ6Z08/p1629730142416200?thread_ts=1629493238.390800&cid=C01T8EZ6Z08


Approved-by: Oliver Harrison
Approved-by: Max Nicholson
Approved-by: Enrique M Figuerola Gomez