Merged in UD-900-remove-filters-on-child-nodes-2 (pull request #488)
UD-900 Optimize loader queries

* UD-900 Optimize from loader queries

This commit should not cause any external behaviour change.
* Remove filters that have already been applied to the parent and therefore cannot change the results
* If a JOIN is needed for data regardless of filtering, do it regardless of filters (to avoid typeOrm making a second query for the data)

The remaining filters on the UserConnection queries do change the results in some cases. This is because the functions return entities that have many-to-one or many-to-many relationships with User ID. Even if a property of the join to those entities has been checked by the parent, that only gurantees that at least one of the entities should be returned. We must filter the child's joins to find out which one(s)

orgsForUsers: It joins to organizations from user IDs. We need to re-apply filters on organisations and roles.
schoolsForUsers: it joins to schools (and then organizations), we must re-apply filter of schools, school organisations and school roles
rolesForUsers: same as the above two, for each of it school and organization queries


Approved-by: Oliver Harrison
Approved-by: Enrique M Figuerola Gomez