Merged in UD-728-Teacher-can-not-see-some-users-on-Users-page (pull request #527)
UD-728 Teacher can not see some users on Users page

* Show other teachers in queries for classes

* Make nonAdminUserScope queries for users more efficent

Previously these queries did subqueries with  many-to-many joins between classes and users to get distinct lists of students or teachers in a class.
But both user ID and class ID are directly avaible in the membership join tables, we don't need to join to user or class in the subqueries

* test: students/teachers with createClass factory

* Refactor isAdmin test for view_my_class_users_40112 into many smaller tests

To make it clearer which exact scenarios we are try to test

* Fix nonAdminUserScope query if user with view_my_class_users_40112 teaches no classes

If this happens `classesTaught` will be an empty array, which causes the subquery for distinct teachers/students to pass an empty list to the IN operator:
```
    SELECT
        DISTINCT "membership_table"."userUserId" AS "user_id"
    FROM "user_classes_studying_class" "membership_table"
    WHERE "membership_table"."classClassId" IN ()
```
Postgres does not like that.

Now it skips the classesTaught queries and if the user has no other permissions will fall through to only allowing them to see themselves
https://bitbucket.org/calmisland/kidsloop-user-service/src/a59e0bff7258acaae092b08e9a668653a63a2175/src/directives/isAdmin.ts#lines-209


Approved-by: Matthew Carey
Approved-by: Max Nicholson