Merged in UD-1106-depth-limit (pull request #604)
feat(UD-1106): enforce depth limit on graphql queries

* chore(UD-1106): use volume in docker-compose so you don't need to rebuild it all the time

* feat(UD-1106): enforce depth limit on graphql queries

Chose a limit of 10 to allow 3 nested connections + fields on the inner most child
10 is actually too high for our existing fields because the deepest use cases are maybe 4 layers deep:

```
user -> school -> classes -> teachers
```

But that's ok because we don't have any limit at the moment, so 10 is an improvement at least.

Nested connections need a higher limit because they introduce extra layers because of the `edge` and `node` fields.


Approved-by: Oliver Harrison
Approved-by: Max Nicholson