fix: async functions inside foreach
* fix: async functions inside foreach

This was the cause of flakey test with errors about deadlocks in afterEach
when truncating tables

If you do `await` inside an async function in a foreach it doesn't block the loop
so the `foreach` may finish and continue to code after it while the inner function
is still processing for some elements.

https://stackoverflow.com/questions/37576685/using-async-await-with-a-foreach-loop

Also fixed the instance of this in src, as TODO pointed out it needed fixing

I can't see a lint specificly for this, only one for `await` inside all loops
we can't set that to `error` as
1) we do it in many places
2) its a bit too strict, sometimes its ok to do this

So instead we just `warn` on this.


Approved-by: Oliver Harrison
Approved-by: Matthew Carey