fix(AD-1769): validate UTF-8 in file uploads
* fix(AD-1769): validate UTF-8 in file uploads

In production it seems like clients cannot be expected to provided
the encoding of the uploaded file, tested in alpha the encoding
field was empty.

So we check the file is valid utf-8 and if so, assume that's the
intended encoding.

`csv-parser` already decodes files as UTF-8 internally so for
valid files the behaviour should be unchanged.

We strip BOM bytes, as if these are present its a strong indicator
the file is intended to be UTF-8 encoded.


Approved-by: Matthew Entwistle
Approved-by: Matthew Carey