fix(AD-1810): undefined authentication tokens
* fix(AD-1810): handle undefined and return 401 codes

The backend CMS doesn't supply an auth token and we need to
support that.
Previously we were applying a revert to allow that on production
branches.

But this led to us developing code that didn't account for that
edge case. So this change applies that change to master directly.

(which will also fix the issue of no authentcated playground in prod
https://kidsloop.slack.com/archives/C02LP6QAQGZ/p1639068719138000)

Because the CMS use-case is about providing no authentication
token at all, we can still enforce errors for bad tokens.

Previously authentication errors thrown during context creation
would give a status code of 400, but for a future change FE need
a 401 error for bad tokens.

See https://bitbucket.org/calmisland/kidsloop-hub-frontend/pull-requests/513

To support that, we move token checking into a middleware where
we can control a 401 status code.

There was already a partial authcheck done as a middleware,
I'm not sure why that existed..

As authentication is enforced in the express layer, authentication
tests are moved from intergration to acceptance.

* fix: remove blacklist issuer logic

redundant because these issuers aren't in the issuer map we check anyway

* fix: unit tests

* fix: make id of token optional

Approved-by: Matthew Entwistle
Approved-by: Oliver Harrison
Approved-by: Toghrul Taghiyev