chore: make it possible to batch process csv rows
The current CSV framework assumes each row can be proccessed in isolation.
But we want to batch DB calls for the users CSV to imrpove performance

This changes the framework to expect functions that receive multiple
rows at a time.

`processUserFromCSVRows` will take advantage of this by having
return entites to save, collecting them and saving them all at once.

However this will require refactoring of how `processUserFromCSVRow`
does validation. So it'll be done in a follow up PR.

For other CSV mutations, use a dummy wrapper that doesn't take
advantage of this but allows the old functions to work with the new
callback type.

50 for the batch size is fairly arbitary
* partly based on our mutation input length, if we 50 at a time there, we can handle it here
* partly based on max CSV file size being 50*1024, implies 50 length is expected