chore(AD-2355): make deprecated field logs valid json
I was trying to write a new relic query that used message fields
as facets, like:

```
SELECT
    count(*)
FROM Log
FACET entity.name, message.deprecatedField
WHERE message LIKE '%deprecated%' AND entity.name LIKE 'user-service%'
```

However that's not possible because new relic is not parsing
our message field.

[This](https://docs.newrelic.com/docs/logs/ui-data/parsing/) page says

```
If the content of the message field is JSON, it will be parsed automatically.
```

So by making it JSON, we'll be able to make nicer charts from our
deprecation logs.

Example of a current message that is not JSON:

```
{
  operationType: 'query',
  operationName: 'studentsByOrganization',
  deprecatedField: 'user_name',
  originURL: 'https://cms.alpha.kidsloop.net'
}
```