chore: remove createMigrationTestConnection
Mainly because it isn't needed, and having it means we need to
duplicate the connection config between 2 functions

Doing this exposed a bug in our migration tests.
Somes we ran

```
   const pendingMigrations = await baseConnection.showMigrations()
   expect(pendingMigrations).to.eq(false)
```

To check that all migrations have run as part of the test.

However, where `baseConnection` was from `createTestConnection`
it would always return false, even if not all migrations had been run
Because `createTestConnection` did not supply a `migrations` key to
the connection it creates - meaning that typeORM cannot see any
migration files for that connection
(and so thinks there are none to run)